+++ examples/ask_age.js
@@ -1,21 +1,18 @@
-var babble = require('../index'),
-    babbler = babble.babbler,
-    reply = babble.reply,
-    run = babble.run;
+var babble = require('../index');
 
-var emma = babbler('emma').subscribe(),
-    jack = babbler('jack').subscribe();
+var emma = babble.babbler('emma').subscribe(),
+    jack = babble.babbler('jack').subscribe();
 
-emma.listen('ask age', reply(function () {
+emma.listen('ask age', babble.reply(function () {
   return 25;
 }));
 
-emma.listen('tell age', run(function (age) {
+emma.listen('tell age', babble.run(function (age) {
   console.log(this.from + ' is ' +  age + ' years old');
 }));
 
 jack.tell('emma', 'tell age', 27);
 
-jack.ask('emma', 'ask age', run(function (age) {
+jack.ask('emma', 'ask age', babble.run(function (age) {
   console.log(this.from + ' is ' + age + ' years old');
 }));