+++ HISTORY.md
@@ -6,6 +6,7 @@ https://github.com/josdejong/babble
 
 - Implemented customizable pubsub system.
 - Implemented support for pubnub.
+- Implemented browser support.
 
 
 ## 2013-12-31, version 0.2.0